6/6 Assignment:  ParseHub Dataset
The program worked almost perfectly.  It iterated through each value on each page, and scraped the title, URL, date, and type as intended.  Some issues I ran into were that the program repeated each entry one time for every entry on the page, indicating some loop issue.  Another issue was that the labels for each column were incorrect.  This might be due to an issue with naming the elements within the script itself.  Overall, I believe I attained a rudimentary understanding of the program, and that I could achieve proficiency fairly quickly if I spend a few more hours toying around with it.